backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
httpx==0.28.1
anthropic==0.43.0
tavily-python==0.5.0
python-dotenv==1.0.1
